Fully functional
Tested for unique values of classes collection as well as for each individual
class.

In all cases, either a dictionary or frozenset are used for returned objects.
A lookup from a set/frozenset is much faster than a lookup from a list/tuple.
Lookup for dictionary is equivalent to set/frozenset.
http://juliank.wordpress.com/2008/04/29/python-speed-x-in-list-vs-x-in-set/

I have a slight concern that class inheriting dict(landCoverClasses
object) might be slower than for a dict used directly...it might be worth a test
at some point.