Made bug fixes in the metric.py script to only delete the tabAreaTable if it exists and, in the LCSOP metric, to delete the clipGrid if created and to reset the original geoprocessing environment at completion. Corrected bug in CEAM metric that wouldn't allow you to save the intermediate grid when the clipGrid to smallest recommended size option was selected. Cleaned up user messages throughout the script to better state what ATtILA is doing while running and to improve readability of the script for collaborators. Added code to start the calculation of the metrics in the new Patch Metrics script tool.
In raster.py, made cosmetic changes to improve readability of the script, and added a new function to generate the patch raster for the Patch Metrics tool. The patch raster will be coded  with a unique number starting at 1 for each individual land cover patch for the selected land cover class, -9999 for any excluded values in the land cover grid, and 0 for any other land cover in the grid. No Data areas in the input land cover grid are preserved.