A number of changes to improve and debug the Patch Metrics and MDCP tools
toolvalidator __init__.py: added ability to rename default field for the PLGP metric using the lcc xml file

attila2__Don.tbx: added a patch metrics tool that doesn't require a search radius for the MDCP metrics. Left the tool with the search radius input box in place for the meantime in case I want to go that route at a later date.

calculate.py: 1)the MDCP tool will now only report out the Patches with Neighbors (PWN) and Patches without Neighbors (PWON) fields if the QAFields option is checked in the tool's Additional Options section, 2) debugged the Patch Metrics tool to handle cases where no land cover is within the reporting unit, where no patches exist within the reporting unit , and tried to speed things up by shifting the analysis extent to each reporting unit before performing the tabulate area command.

LCCSchema_v2.xsd: added code for the schema to accept the pmfield information in the lcc xml file.

metrics.py: 1) debugged the patch metrics output table to only add the MDCP field(s) if the MDCP checkbox is toggled, 2) renamed a number of the intermediate files to more closely follow the conventions set in other ATtILA tools, and 3) set up the patch metric to return a dictionary of results so that the MDCP script could cross reference answers to see if the number of patches determined by the two techniques were the same.

metricConstants.py: altered the constants for the patch metrics tool to improve the intermediate file names

vector.py: commented out a number of original iterations of tool functions. They were retained as references as the final script functions were improved and debugged. The plans are to delete the commented out functions after they are recorded in a commit operation to the github repository and can be referenced again if needed from the commit logs.

PatchMetricsSearch.txt: a text file copy of the Patch Metrics tool validation script.